Create a new component task-list. Move the logic of adding/removing tasks and rendering of tasks UI to the new component
