+++ src/app/app.component.css
@@ -1,3 +1,1 @@
-.is-done {
-  text-decoration: line-through;
-}
+
