+++ src/app/app.component.ts
@@ -8,20 +8,4 @@ import { Task } from './Task';
 })
 export class AppComponent {
   title = 'TODOs App';
-  tasks: Task[] = [];
-
-  addTask(taskRef: any) {
-    this.tasks.push(new Task(taskRef.value, false));
-    taskRef.value = '';
-    taskRef.focus();
-  }
-
-  removeTask(task: Task) {
-    const response = confirm(
-      `Are you sure you want to remove the task "${task.title}"?`
-    );
-    if (response) {
-      this.tasks = this.tasks.filter((eachTask) => eachTask !== task);
-    }
-  }
 }
