+++ src/app/task-list/task-list.component.css
@@ -1,0 +1,3 @@
+.is-done {
+  text-decoration: line-through;
+}
